Add existence check
Former-commit-id: 4be708d336132f4b576212e55f6644c381e678c3

+++ org.moflon.ide.editor/src/org/moflon/ide/texteditor/helpers/MarkerHelper.java
@@ -36,7 +36,8 @@ public class MarkerHelper {
 	 */
 	public static void removeMarkers(IResource resource) {
 		try {
-			resource.deleteMarkers(WorkspaceHelper.MOFLON_PROBLEM_MARKER_ID, true, IResource.DEPTH_INFINITE);
+			if (resource.exists())
+				resource.deleteMarkers(WorkspaceHelper.MOFLON_PROBLEM_MARKER_ID, true, IResource.DEPTH_INFINITE);
 		} catch (CoreException e) {
 			e.printStackTrace();
 		}
